split the playlist sidebar to models
